split languae_analysis module
